+++ src/xulbux/string.py
@@ -39,7 +39,7 @@ class String:
 
     @classmethod
     def escape(cls, string: str, str_quotes: Optional[Literal["'", '"']] = None) -> str:
-        """Escapes Python's special characters (e.g. `\\n`, `\\t`, ...) and quotes inside the string.\n
+        """Escapes Python's special characters (e.g. `\\n`, `\\t`, …) and quotes inside the string.\n
         --------------------------------------------------------------------------------------------------------
         - `string` -⠀the string to escape
         - `str_quotes` -⠀the type of quotes the string will be put inside of (or None to not escape quotes)<br>